requirements.txt
Flask
Flask-CORS
Flask-Bcrypt
python-dotenv
Flask-SocketIO
eventlet